+++ src/main/java/com/ohacd/matchbox/game/utils/ProjectStatus.java
@@ -1,0 +1,20 @@
+package com.ohacd.matchbox.game.utils;
+
+public enum ProjectStatus {
+    STABLE("Stable"),
+    DEVELOPMENT("Development"),
+    BETA("Beta"),
+    ALPHA("Alpha"),
+    RELEASE_CANDIDATE("Release Candidate"),
+    RELEASE("Release");
+
+    private final String displayName;
+
+    ProjectStatus(String displayName) {
+        this.displayName = displayName;
+    }
+
+    public String getDisplayName() {
+        return displayName;
+    }
+}