+++ src/main/java/com/ohacd/matchbox/game/utils/listeners/HitRevealListener.java
@@ -1,8 +1,11 @@
-package com.ohacd.matchbox.game.utils;
+package com.ohacd.matchbox.game.utils.listeners;
 
 import com.ohacd.matchbox.game.GameManager;
 import com.ohacd.matchbox.game.SessionGameContext;
 import com.ohacd.matchbox.game.hologram.HologramManager;
+import com.ohacd.matchbox.game.utils.GamePhase;
+import com.ohacd.matchbox.game.utils.Managers.InventoryManager;
+
 import org.bukkit.entity.Arrow;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;