+++ src/main/java/com/ohacd/matchbox/game/utils/listeners/PlayerJoinListener.java
@@ -1,6 +1,7 @@
-package com.ohacd.matchbox.game.utils;
+package com.ohacd.matchbox.game.utils.listeners;
 
 import com.ohacd.matchbox.Matchbox;
+import com.ohacd.matchbox.game.utils.CheckProjectVersion;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.title.Title;
 import org.bukkit.entity.Player;
@@ -15,9 +16,15 @@ import java.time.Duration;
  */
 public class PlayerJoinListener implements Listener {
     private final Matchbox plugin;
-    
-    public PlayerJoinListener(Matchbox plugin) {
+    private final CheckProjectVersion versionChecker;
+    // Keeps track of the version and the current status of the project
+    private final String currentVersion = Matchbox.getInstance().getCurrentVersion();
+    private final String projectStatus = Matchbox.getInstance().getProjectStatus();
+    private final String updateName = Matchbox.getInstance().getUpdateName();
+
+    public PlayerJoinListener(Matchbox plugin, CheckProjectVersion versionChecker) {
         this.plugin = plugin;
+        this.versionChecker = versionChecker;
     }
     
     @EventHandler
@@ -52,8 +59,8 @@ public class PlayerJoinListener implements Listener {
                     player.sendMessage("§7players must work together to identify the §cSpark §7(impostor)");
                     player.sendMessage("§7while the Spark tries to eliminate everyone.");
                     player.sendMessage("");
-                    player.sendMessage("§7Current Version: §e0.9.1 §7(Config and QOL Update)");
-                    player.sendMessage("§7Status: §aStable §7- Ready for gameplay");
+                    player.sendMessage("§7Current Version: §e" + currentVersion + " §7(" + updateName + ")");
+                    player.sendMessage("§7Status: §d" + projectStatus); // " §7- Ready for gameplay"
                     player.sendMessage("");
                     player.sendMessage("§7Found a bug or have suggestions?");
                     player.sendMessage("§7Join our Discord: §9§nhttps://discord.gg/BTDP3APfq8");
@@ -61,6 +68,29 @@ public class PlayerJoinListener implements Listener {
                 }
             },
             40L // 2 seconds delay (20 ticks = 1 second)
+        );
+
+        // Checks the version that the player is running against the latest project version and notify the player
+        plugin.getServer().getScheduler().runTaskLater(
+                plugin,
+                () -> {
+                    versionChecker.checkLatestVersion(latest -> {
+                        if (latest == null) {
+                            player.sendMessage("§cCould not fetch version info.");
+                            return;
+                        }
+
+                        if (currentVersion.equalsIgnoreCase(latest)) {
+                            player.sendMessage("§aYou're all good and running the latest version! (" + latest + ")");
+                        } else {
+                            player.sendMessage("§eA newer version available: §b" + latest);
+                            player.sendMessage("§eYou are currently on: §c" + currentVersion);
+                            player.sendMessage("§eYou can download the latest version here:");
+                            player.sendMessage("§9§nhttps://modrinth.com/plugin/matchboxplugin");
+                        }
+                    });
+                },
+                50L
         );
     }
 }