+++ src/main/java/com/ohacd/matchbox/game/utils/listeners/VoteItemListener.java
@@ -1,7 +1,10 @@
-package com.ohacd.matchbox.game.utils;
+package com.ohacd.matchbox.game.utils.listeners;
 
 import com.ohacd.matchbox.game.GameManager;
 import com.ohacd.matchbox.game.SessionGameContext;
+import com.ohacd.matchbox.game.utils.GamePhase;
+import com.ohacd.matchbox.game.utils.Managers.InventoryManager;
+
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;